feat: Create fixtures.py and move user, authenticated_user, and api_client fixtures
- Creates `tests/fixtures.py` to house shared fixtures.
- Moves `user`, `authenticated_user`, and `api_client` fixtures to `tests/fixtures.py`.
- Refactors tests to use the shared fixtures.
- Fixes test setup issues.
- Addresses PR comments.

Fixes #30
