+++ tests/conftest.py
@@ -1,4 +1,9 @@
 import pytest
+from django.contrib.auth import get_user_model
+from rest_framework.authtoken.models import Token
+from rest_framework.test import APIClient
+
+from tests.factories import UserFactory
 
 """
 Global pytest configuration file.
@@ -44,4 +49,35 @@ def _override_celery_settings_for_tests(settings):
 def aaa_db(db):
     pass
 
-pytest_plugins = "tests.fixtures"
+
+User = get_user_model()
+
+
+@pytest.fixture
+def user(db):
+    """Provides a user instance."""
+    return UserFactory()
+
+
+@pytest.fixture
+def authenticated_user(client, user):
+    """
+    Provides an authenticated user by force-logging them in.
+
+    This fixture uses `client.force_login()` for efficiency, as it directly
+    logs in a user without going through the session and password validation
+    process. This is suitable for most tests that simply require an
+    authenticated user context.
+    """
+    client.force_login(user)
+    return user
+
+
+@pytest.fixture
+def api_client(db):
+    """Provides an authenticated API client."""
+    user = UserFactory()
+    token = Token.objects.create(user=user)
+    client = APIClient()
+    client.credentials(HTTP_AUTHORIZATION=f"Token {token.key}")
+    return client
